fix: Correct the true issue for sorting icons missing in DataTables
I thought there were no icons at all, as described in #79, some problem
with loading the images or CSS rules. Turns out the problem is we have
the icons locally, packed with the extension. And in order to load
static files, you need the absolute path that depends on the extension
ID. Moreover, it depends on the browser itself. For instance, chrome
uses the `chrome-extension://` protocol, while firefox uses
`moz-extension://` protocol.

The the datatables.min.css you can see URLs like
`chrome-extension://__MSG_@@extension_id__`, a trick to represent the
absolute path that can be used in CSS. Note that in JS there is a proper
API for that. Firefox does not support this link in CSS, thus.. no
icons! Fortunatelly, it works if I change the protocol for
`moz-extension`.

The final solution was to override the CSS rules with a
`background-image` with two URLs. Technically, this is used to overlap
multiple background images. But I know that one of them will fail
accordingly with the browser you running. Thus.. it works and prevents a
JS approach to set the URLs dynamically.

I also updated the icons to be more visible.