fix: Updates extractors paths for requesting HTML on unit tests
The updatejQueryContext is updated to prepend a localhost:3000 address
on all URL requests. Besides, all extractors requeted paths are updated.
Before, they started on /test/pages/<...>. Since this is only used for
HTML pages under /test/pages/, I think we could remove that part. The
server can actually have the root on that directory, recuding the length
of paths in the code