fix: Improve 'one row' validation
The previous heuristic was to get all <tr> of a table and if the length
is >= 2, then it is valid. The reasoning was one <tr> is for the header,
thus we need two extra <tr> that should be the table body.

Since there is a method that gathers the table header rows, the new
approach is smarter because it does not assume a single header row.
Computing the diff <all rows> - <header rows> gives the data/body rows,
which then is used for validation