enhance: Different approach to DataTables allowing full recovery on failure
Before it was like:
- Clone original table DOM
- Try to apply DataTable
- If fails, try to restaure the original DOM, cloned in first step

The recovery is not that easy because the DataTable can fail at very
different stages. Perhaps, it just added the wrapper containers for the
buttons (csv, excel, ...) and search input. But I found cases where it
goes further and inserts sorting event listeners and the buttons in some
columns. There are also scenarios where it fails in very early stages,
adding just a few classes in 'class' attribute. Thus, it becomes
difficult to ensure a clean DOM like the original one.

With this commit, the procedure is:
- Clone original table
- Try to pre-process and apply DataTable in the cloned version
- If OK, then replace the original table in the DOM with the new DOM
  containing filter/search components, the DataTable itself, etc.

Advantages for the new approach are:
- 100% clean DOM upon failure
- More performant. The DOM in the webpage is manipulated once per table
  on success. In case of failure, theres no change at all saving
rendering/parsing cycles.