fix: Handle some issues with infinity scroll after DataTable extractor changes
- Main issue was infinity scroll querying DOM elements at the
  constructor before the DataTable extractor completing its job.
Therefore, the DOM would modify and the refereces in Infinite Scroll
would be invalid
- Another issue was the loading gif missing after the first load. The
  reason is when the DataTable is updated with new data, the DOM is
reconstructured and the <img> node for the GIF is destroied. The fix is
to attach this element in the DOM everytime.
- This also changes how data is added to an existing table. There is no
  need to destroy the previous table. The API allows to pass the new
elements and then redraw the table. Besides, you can pass HTML elements
as input (i.e. <tr>)