Add a logging mechanism that can be dissabled in production (#103)
* chore: Add 'preprocess' in gulp

It allows to change JS/HTML/... files depending on environment
variables. It is similar to macros in C where you can inject code if
some environment variables are set. Docs: https://www.npmjs.com/package/gulp-preprocess

The preprocess will be used to enabled/disable a the logger according to
the development and production modes. See #102

* feat(logger): Add a toggling logger mechanism for dev/prod environments