enhance(EventExtractor): Support nested objects for validating parameter objects at runtime
The function `_validateEventParams` receives an object that should have
all properties declared in the extractor `structure.properties`. It is
used to format strings such as event title, description, etc. The
properties may declared nested objects when the name is like
`room.name`, meaning the `room` is an object with at least the property
`name`. Validating nested objects is now supported.