fix(EventExtractor): Properly escape all newlines before using `eval` to format strings
Previous implementation would only escape newlines if the format-strings
are in HTML. However, it would only escape the first occurence, as
`replace` stops on the first occurance. The alternative `replaceAll` is
still not widely available, therefore we keep `replace` but use a regex
pattern with the global (g) flag. It also escapes newlines for plain text string-formats.

Escapping is necessary when the string-format looks like this: `hello
'${p1}\n${p2}'`. As the `\n` is not escaped and it is between single
quotes, `'`, it the eval will fail as single and double quotes do not
work with multi-line strings.