fix(SingleExam): Handle missing parameters
Previous approach assumed that the data is always present and in the
same order. However, that is no longer true.
- The list of teachers/supervisors is optional
- Sometimes there is an observations field in the middle that affects
  the index of the following rows. Thus assuming the list of rooms is on
index `i` is not a safe approach.

The fix consists in searching for rows by label, assuming the labels are
always the same. It also assumes that at least the list of supervisors
is optional.